# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MoneyQuest Personal Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: Get all transactions
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionInput"
      responses:
        "200":
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /subscriptions:
    get:
      operationId: getSubscriptions
      tags: [subscriptions]
      summary: Get all subscriptions
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionInput"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionInput"
      responses:
        "200":
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Delete a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /insights/monthly:
    get:
      operationId: getMonthlyInsights
      tags: [insights]
      summary: Get monthly breakdown
      parameters:
        - name: months
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Monthly data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyData"
  /insights/categories:
    get:
      operationId: getCategoryInsights
      tags: [insights]
      summary: Get spending by category
      parameters:
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryData"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        amount:
          type: number
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        description:
          type: string
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        date:
          type: string
        isRecurring:
          type: boolean
        recurringPeriod:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - amount
        - category
        - description
        - date
        - isRecurring
        - createdAt
    CreateTransactionInput:
      type: object
      properties:
        type:
          type: string
        amount:
          type: number
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        description:
          type: string
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        date:
          type: string
        isRecurring:
          type: boolean
        recurringPeriod:
          type: ["string", "null"]
      required:
        - type
        - amount
        - category
        - description
        - date
    UpdateTransactionInput:
      type: object
      properties:
        type:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        category:
          type: ["string", "null"]
        subcategory:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        date:
          type: ["string", "null"]
        isRecurring:
          type: ["boolean", "null"]
        recurringPeriod:
          type: ["string", "null"]
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        amount:
          type: number
        billingCycle:
          type: string
        nextChargeDate:
          type: string
        category:
          type: string
        isEssential:
          type: boolean
        annualCost:
          type: number
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - amount
        - billingCycle
        - nextChargeDate
        - category
        - isEssential
        - annualCost
        - createdAt
    CreateSubscriptionInput:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        billingCycle:
          type: string
        nextChargeDate:
          type: string
        category:
          type: string
        isEssential:
          type: boolean
      required:
        - name
        - amount
        - billingCycle
        - nextChargeDate
        - category
    UpdateSubscriptionInput:
      type: object
      properties:
        name:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        billingCycle:
          type: ["string", "null"]
        nextChargeDate:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        isEssential:
          type: ["boolean", "null"]
...
    MonthlyData:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        expenses:
          type: number
        savings:
          type: number
      required:
        - month
        - income
        - expenses
        - savings
    CategoryData:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
        transactionCount:
          type: integer
      required:
        - category
        - amount
        - percentage
        - transactionCount
...