# lib/api-zod/src/generated/types/updateTransactionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MoneyQuest Personal Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTransactionInput {
  /** @nullable */
  type?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  subcategory?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  /** @nullable */
  date?: string | null;
  /** @nullable */
  isRecurring?: boolean | null;
  /** @nullable */
  recurringPeriod?: string | null;
}